+++ src/XsDupFinder.Lib/Output/Renderer/IRender.cs
@@ -1,12 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using XsDupFinder.Lib.Common;
 using XsDupFinder.Lib.Finder;
 
-namespace XsDupFinder.Lib.Output
+namespace XsDupFinder.Lib.Output.Renderer
 {
     interface IRender
     {
